feat(web): add persist tab dialog
