+++ web/src/app/dashboard/workspace/[workspace]/@modal/(.)task/[task]/layout.tsx
@@ -1,10 +1,9 @@
 import React, { type JSX } from "react";
-import TaskDialogProvider from "./_providers/task-dialog";
 
-export default function TaskTabLayout({
-  children,
-}: {
+type Props = Readonly<{
   children: React.ReactNode;
-}): JSX.Element {
-  return <TaskDialogProvider>{children}</TaskDialogProvider>;
+}>;
+
+export default function TaskTabLayout({ children }: Props): JSX.Element {
+  return <>{children}</>;
 }
